Update spec coverage for insurance helpers in Health Care Application
